feat: move blocks for cases to separate folder
