+++ src/components/pages/case/bg-decor/bg-decor.jsx
