+++ src/components/pages/case/bg-decor/index.js
