+++ src/components/pages/case/case-cta/case-cta.jsx
@@ -10,8 +10,8 @@ import blueGlow from './images/blue-glow.png';
 import greenGlowMobile from './images/green-glow-mobile.png';
 import greenGlow from './images/green-glow.png';
 
-const Cta = ({ text, button }) => (
-  <div className="relative mt-10 w-full overflow-hidden rounded-lg bg-gray-new-8 px-7 py-6 xl:mt-9 lg:mt-8 lg:pr-10 sm:mt-6 sm:p-6">
+const CaseCta = ({ text, buttonText, buttonUrl }) => (
+  <div className="not-prose relative mt-10 w-full overflow-hidden rounded-lg bg-gray-new-8 px-7 py-6 xl:mt-9 lg:mt-8 lg:pr-10 sm:mt-6 sm:p-6">
     <div className="relative z-10 flex items-center justify-between sm:flex-col sm:gap-[18px]">
       <h3 className="mb-1 max-w-sm text-2xl font-medium leading-dense tracking-tighter sm:mb-1 sm:text-center sm:text-xl">
         {text}
@@ -19,9 +19,9 @@ const Cta = ({ text, button }) => (
       <Button
         className="h-10 px-7 text-base !font-semibold tracking-tighter lg:text-sm"
         theme="primary"
-        to={button.url}
+        to={buttonUrl}
       >
-        {button.title}
+        {buttonText}
       </Button>
     </div>
     <BgDecor hasBorder hasNoise hasPattern>
@@ -57,12 +57,10 @@ const Cta = ({ text, button }) => (
   </div>
 );
 
-Cta.propTypes = {
+CaseCta.propTypes = {
   text: PropTypes.string.isRequired,
-  button: PropTypes.shape({
-    title: PropTypes.string.isRequired,
-    url: PropTypes.string.isRequired,
-  }).isRequired,
+  buttonText: PropTypes.string.isRequired,
+  buttonUrl: PropTypes.string.isRequired,
 };
 
-export default Cta;
+export default CaseCta;
