+++ src/components/pages/case/case-cta/images/blue-glow.png
