+++ src/components/pages/case/case-cta/images/green-glow-mobile.png
