+++ src/components/pages/case/case-cta/index.js
@@ -1,0 +1,3 @@
+import CaseCta from './case-cta';
+
+export default CaseCta;
