--- src/components/pages/variable/cta/index.js
@@ -1,3 +1,0 @@
-import Cta from './cta';
-
-export default Cta;
