+++ src/components/pages/variable/efficiency/efficiency.jsx
@@ -1,5 +1,6 @@
 import { PropTypes } from 'prop-types';
 
+import Testimonial from 'components/pages/case/testimonial';
 import autoscaleIcon from 'icons/variable/autoscale.svg';
 import costIcon from 'icons/variable/cost.svg';
 import databaseIcon from 'icons/variable/database.svg';
@@ -10,7 +11,6 @@ import pieralbertoColonboAvatar from 'images/pages/variable-load/testimonials/pi
 
 import List from '../list';
 import Section from '../section';
-import Testimonial from '../testimonial';
 
 const items = [
   {
