+++ src/components/pages/variable/unique/unique.jsx
@@ -1,11 +1,11 @@
 import { PropTypes } from 'prop-types';
 
+import Testimonial from 'components/pages/case/testimonial';
 import LINKS from 'constants/links';
 import andreaLevingeAvatar from 'images/pages/variable-load/testimonials/andrea-levinge.jpg';
 
 import List from '../list';
 import Section from '../section';
-import Testimonial from '../testimonial';
 
 const items = [
   {
